extract shared helpers: buildActionFromFlags, interactive-helpers (PlanInfo, getAvailableConditionTypes, getActionLabel, getOperatorDisplayName, validateConditionValue); remove --ai from agent error; remove dead outputAgentError from AI handler
